Hey,
This should be it. For now... I applied most of the remarks you told me to. Only the seemingly floating squares... well let's just say they add to the mysteriousness of the page. I can't delete them. I don't think my site would look without them. We'll just see what the users think of it shall we? I can ask them abou the squares once it's released. So, all there is now is the fat vertical bar in NS 4.76. I also don't know what it looks like in Opera... and UnConeD, does it show a bit OK in IE? I polished the navigation box' algorithm a bit. Even more stable now.

I do have one remark Dries. When submitting something you don't have to choose story or book anymore but you have to select one of the categories. Is this OK? But wouldn't it be more logical to choose what type of thing you are going to submit. I don't think you will be placing an Announcement in the book. By the way how do you add a page to the book now? I can't find it... Or do I have to create book and story myself... Ahh heck Dries it's just broken OK? ;)

For the rest. Oh yes, I am not going to work with the categories. The squares in front of a $story->title represent the topic the title is under (e.g.: linux, science, Media, Announcements) But not the category (I see them as a some sort of sub-folder in each topic) I haven't found a good way to display it yet. I still have to find how to import it in the lay-out so it wouldn't mess up. I already liked the fact that I don't have text-links for the former categories anymore... We'll see what happens...

Jeroen.