Added poll.module, a poll module :)
It's mainly intended as a sideblock-element as there's no real use in having them show up on the mainpage. There's no real poll-browsing system for now either.

I still need to write documentation, and I'm not yet satisfied with the node-output (non-sideblock). I should be able to improve it once the $theme->nodebox ideas get finalized.